requirements.txt
Flask==1.1.1
Flask-Login==0.5.0
Werkzeug==1.0.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
WTForms==2.2.1

requests==2.22.0

SQLAlchemy==1.3.13
psycopg2-binary==2.8.5

vk==2.0.2
discord.py==1.3.2

Pillow==5.4.1
wikipedia==1.4.0